Multimedia, views & components organized properly. Files renamed. Comments dropped.
