+++ client/src/components/navBar/Navbar.css
@@ -3,7 +3,7 @@
 	flex-direction: column;
 	background-color: black;
 	padding-right: 15px;
-	background-image: url('./brick-wall-1.jpg');
+	background-image: url('../../multimedia/navbar-bg.jpg');
 	padding-bottom: 5px;
 }
