+++ client/src/multimedia/loading.gif
