+++ client/src/views/gameDetails/gameDetails.css
@@ -8,7 +8,7 @@
 	flex-direction: column;
 	justify-content: space-around;
 	align-items: center;
-	background-image: url('./depositphotos_63921325-stock-photo-arcade-machine-opposing-duel.jpg');
+	background-image: url('../../multimedia/gameDetails-bg.jpg');
 	background-repeat: no-repeat;
 	background-size: cover;
 	background-position: center center;
