+++ client/src/views/gameDetails/index.js
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useParams } from 'react-router-dom';
 import { getVideogameDetails } from '../../actions/actions';
-import './GameDetails.css';
+import './gameDetails.css';
 
 const GameDetails = () => {
 	const dispatch = useDispatch();
