+++ client/src/views/home/home.css
@@ -1,5 +1,5 @@
 .home {
-  background-image: url("./neon-arcade-machines-2.png");
+  background-image: url("../../multimedia/home-bg.png");
   background-repeat: no-repeat;
   background-size: cover;
   background-position: top;
