+++ client/src/views/landingPage/index.js
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import "./LandingPage.css";
+import "./landingPage.css";
 
 const LandingPage = () => {
   return (
