+++ client/src/views/landingPage/landingPage.css
@@ -1,5 +1,5 @@
 .container {
-  background-image: url("./texto-estilo-letreros-neon-arcade-night_118419-1578.jpg");
+  background-image: url("../../multimedia/landingPage-bg.jpg");
   background-repeat: no-repeat;
   background-size: cover;
   background-position: center center;
